requirements.txt
numpy==1.23.5
pandas==1.5.3
matplotlib==3.7.1
seaborn==0.12.2
scikit-learn==1.2.2
xgboost==1.7.6
openpyxl==3.0.10
joblib==1.2.0
PyYAML==6.0